# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for LLM trace logging, testing, and regression detection",
        "title": "Regrada API",
        "contact": {
            "name": "API Support",
            "email": "support@regrada.ai"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/api-keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get all API keys for your organization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "List API keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "api_keys": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/types.APIKeyResponse"
                                    }
                                },
                                "count": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create a new API key for your organization. The secret is only returned once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api-keys"
                ],
                "summary": "Create a new API key",
                "parameters": [
                    {
                        "description": "API key details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.CreateAPIKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.APIKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "key_prefix": {
                    "type": "string",
                    "example": "rg_live_abc123"
                },
                "last_used_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Production API Key"
                },
                "rate_limit_rpm": {
                    "type": "integer",
                    "example": 100
                },
                "revoked_at": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "traces:write",
                        "tests:write"
                    ]
                },
                "tier": {
                    "type": "string",
                    "example": "standard"
                }
            }
        },
        "types.CreateAPIKeyRequest": {
            "type": "object",
            "required": [
                "name",
                "tier"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Production API Key"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "traces:write",
                        "tests:write"
                    ]
                },
                "tier": {
                    "type": "string",
                    "example": "standard"
                }
            }
        },
        "types.CreateAPIKeyResponse": {
            "type": "object",
            "properties": {
                "api_key": {
                    "$ref": "#/definitions/types.APIKeyResponse"
                },
                "secret": {
                    "type": "string",
                    "example": "rg_live_abcdefghijklmnopqrstuvwxyz123456"
                }
            }
        },
...
        "types.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/types.Error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}